Move object modules to the obj folder to stop shadowing of python keyword.
